Remove redundant AbstractSimpleSecurityTransportTestCase
We only have the Netty based security transport now, no need
for an abstract class like that.
